📦 NEW: Add filter transition
