+++ frontend/src/components/courses/index.jsx
@@ -1,5 +1,4 @@
 export { default as CourseList } from 'components/courses/CourseList'
-export { default as CourseSearchbar } from 'components/courses/CourseSearchbar'
 export { default as PageNo } from 'components/courses/PageNo'
 export { default as CourseItem } from 'components/courses/item'
 export { default as FiltersBody } from 'components/courses/FiltersBody'
